Add root HTML template
Add a root HTML template that will be used as the template for most
pages. It includes the page metadata, header, and footer and allows the
page content to be specified by setting the $page placeholder.

The syntax for including files and variables is only a vague idea at the
moment. `@x` will include the template `x`, and `$y` will be substituted
with the value of variable `y`. The identifier can be enclosed in curly
braces in case it is immediately followed by valid identifier characters
(like `${foo}bar`).